refactoring binary, new tests
