+++ readme.html
@@ -983,6 +983,7 @@ body {
 	"autoSplit": false,
 	"sanitize": false,
 	"noEmbed": false,
+	"edit": true,
 	"title": "README"
 }
